+++ TrueCraft.Client/Rendering/Blocks/LadderRenderer.cs
@@ -25,7 +25,7 @@ namespace TrueCraft.Client.Rendering.Blocks
         };
 
         public override Vertex[] Render(BlockDescriptor descriptor, Vector3 offset,
-             Tuple<int, int> textureMap, int indiciesOffset, out ushort[] indicies)
+             Tuple<int, int> textureMap, int indiciesOffset, out uint[] indicies)
         {
             Vertex[] verticies;
             Vector3 correction;
